+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 CHANGELOG
 =========
 
+
 ### 2017-10-10
  - MIT Licence
  - Callback is no longer a filter
  - Updated README
+ - Function accepts '_route' and '_route_params' as parameters 
  - v2.0.0 New release
 
 ### 2017-08-18